Refactor: api calls refactored into separate components
